+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/rtc/PeerNetworkConfig.kt
@@ -2,13 +2,19 @@ package jp.shiguredo.sora.sdk.channel.rtc
 
 import jp.shiguredo.sora.sdk.channel.option.SoraMediaOption
 import jp.shiguredo.sora.sdk.channel.signaling.message.OfferConfig
+import jp.shiguredo.sora.sdk.util.SoraLogger
 import org.webrtc.CryptoOptions
 import org.webrtc.PeerConnection
 
 class PeerNetworkConfig(
     private val serverConfig: OfferConfig?,
     private val mediaOption: SoraMediaOption,
+    private val insecure: Boolean = false,
 ) {
+    companion object {
+        private val TAG = PeerNetworkConfig::class.simpleName
+    }
+
     fun createConfiguration(): PeerConnection.RTCConfiguration {
         val iceServers = gatherIceServerSetting(serverConfig)
 
@@ -38,16 +44,19 @@ class PeerNetworkConfig(
     private fun gatherIceServerSetting(serverConfig: OfferConfig?): List<PeerConnection.IceServer> {
         val iceServers = mutableListOf<PeerConnection.IceServer>()
         serverConfig?.let {
-            it.iceServers.forEach {
-                val server = it
-                server.urls.forEach {
-                    val url = it
+            it.iceServers.forEach { server ->
+                server.urls.forEach { url ->
                     iceServers.add(
                         PeerConnection.IceServer
                             .builder(url)
                             .setUsername(server.username)
                             .setPassword(server.credential)
-                            .createIceServer(),
+                            .apply {
+                                if (insecure && url.startsWith("turns:")) {
+                                    SoraLogger.w(TAG, "[rtc] insecure is enabled for TURN-TLS: $url")
+                                    setTlsCertPolicy(PeerConnection.TlsCertPolicy.TLS_CERT_POLICY_INSECURE_NO_CHECK)
+                                }
+                            }.createIceServer(),
                     )
                 }
             }